% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_performance_stata_julia_r.R
\name{compare_performance}
\alias{compare_performance}
\title{Function compares performance of Stata, Julia and R boottest implementations
Note that benchmarking Stata through RStata provides some overhead - the STATA
implementation is usually faster than reported here
Stand-alone STATA code can be found below the function}
\usage{
compare_performance()
}
\description{
Function compares performance of Stata, Julia and R boottest implementations
Note that benchmarking Stata through RStata provides some overhead - the STATA
implementation is usually faster than reported here
Stand-alone STATA code can be found below the function
}
