% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_set_nthreads.R
\name{julia_set_ntreads}
\alias{julia_set_ntreads}
\title{Helper function to set the number of threads for use in Julia}
\usage{
julia_set_ntreads()
}
\description{
Helper function to set the number of threads for use in Julia
}
