% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldtest_methods.R
\name{summary.waldtest}
\alias{summary.waldtest}
\title{S3 method to summarize objects of class waldtest}
\usage{
\method{summary}{waldtest}(object, digits = 3, ...)
}
\arguments{
\item{object}{object of type waldtest}

\item{digits}{rounding of output. 3 by default}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Returns result summaries for objects of type waldtest
}
\description{
S3 method to summarize objects of class waldtest
}
