% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldtest_methods.R
\name{tidy.waldtest}
\alias{tidy.waldtest}
\title{S3 method to summarize objects of class waldtest into tidy data.frame}
\usage{
\method{tidy}{waldtest}(object, ...)
}
\arguments{
\item{object}{object of type waldtest}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A tidy data.frame with estimation results for objects of type
waldtest
}
\description{
S3 method to summarize objects of class waldtest into tidy data.frame
}
