% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildboottestjlr_setup.R
\name{wildboottestjlr_setup}
\alias{wildboottestjlr_setup}
\title{install Julia, connect R and Julia, install and pre-compile WildBootTests.jl}
\usage{
wildboottestjlr_setup()
}
\description{
install Julia, connect R and Julia, install and pre-compile WildBootTests.jl
}
