% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildboottestjlr_setup.R
\name{wildboottestjlr_setup}
\alias{wildboottestjlr_setup}
\title{Connect R and Julia via JuliaConnectoR & installs Julia dependency packages (if required)}
\usage{
wildboottestjlr_setup(
  install_julia = FALSE,
  install_wildboottests = TRUE,
  julia_path
)
}
\arguments{
\item{install_julia}{Logical. True by FALSE Should Julia be installed?}

\item{install_wildboottests}{Logical. TRUE by default. Installs WildBootTests.jl}

\item{julia_path}{Character. Path to Julia installation - required if WildBootTests.jl needs to be installed.}
}
\description{
Connect R and Julia via JuliaConnectoR & installs Julia dependency packages (if required)
}
